SKN-361 moving funtion to utils
