+++ src/utils/index.js
@@ -9,9 +9,7 @@ const validateUUID = uuid => {
 
 const getUser = async res => {
     const { username: uid } = res.locals.user;
-    return await Promise.resolve(
-        User.findOne({ where: { uid } })
-    );
+    return await Promise.resolve(User.findOne({ where: { uid } }));
 };
 
 function validateObjectAllocations(obj, periods) {
@@ -152,8 +150,22 @@ function validateAllocations(allocations, level = 0) {
     };
 }
 
+// this function is used to check if a campaign is in flight based on the flight time start and end
+const checkInFlight = ({ currentDate, campaign }) => {
+    const startPeriod = new Date(campaign.flight_time_start);
+    const endPeriod = new Date(campaign.flight_time_end);
+    // Set the endPeriod to the last day of the month
+    endPeriod.setMonth(endPeriod.getMonth() + 1);
+    endPeriod.setDate(0);
+
+    return currentDate >= startPeriod && currentDate <= endPeriod
+        ? true
+        : false;
+};
+
 module.exports = {
     validateUUID,
     getUser,
     validateObjectAllocations,
+    checkInFlight,
 };
